Add hint for node and npm version
Added engines in package.json.
At the moment with this setup we do not enforce specific versions of npm and node. But it is mentioned in the engines section of the package.json.

If it is needed to enforce the npm and node version: see [here](https://stackoverflow.com/questions/29349684/how-can-i-specify-the-required-node-js-version-in-package-json)